actualizacion del firmware de filtros

+++ adc_fir_dac/inc/fir_q31.h
@@ -6,14 +6,14 @@ typedef struct
 	unsigned int last_index;
 	int num_taps;
 	const int * kernel;
-	int * history;
+	float * history;
 }fir_q31_t;
 
-void fir_q31_init(fir_q31_t * f, int * history, const int * kernel, int num_taps);
+void fir_q31_init(fir_q31_t * f, float * history, const int * kernel, int num_taps);
 void fir_q31_put(fir_q31_t * f, int input);
 int fir_q31_get(fir_q31_t * f);
-void set_history(fir_q31_t * f,int * history);
-int get_history(fir_q31_t * f);
+void set_history(fir_q31_t * f,float * history);
+float get_history(fir_q31_t * f);
 
 extern void asm_fir_q31_put(fir_q31_t * f, int input);
 extern int 	asm_fir_q31_get(fir_q31_t * f);
